Gracefully shutdown.
When option `-g` is given to stop, it uses the `HUP` signal to attempt a
clean shutdown. In this case, the supervisor send a `HUP` to all its
children.

This mean that workers needs to trap the `HUP` and implements their own
strategy when it come to have a clean shutdown.

Signed-off-by: chatgris <jboyer@af83.com>